+++ src/main/java/com/jpdr/apps/demo/webflux/stock/exception/product/ProductNotFoundException.java
@@ -2,7 +2,7 @@ package com.jpdr.apps.demo.webflux.stock.exception.product;
 
 public class ProductNotFoundException extends RuntimeException{
   
-  public ProductNotFoundException(Integer productId, Throwable ex){
+  public ProductNotFoundException(Long productId, Throwable ex){
     super("The product " + productId + " wasn't found.", ex);
   }
