+++ src/main/java/com/jpdr/apps/demo/webflux/stock/exception/product/ProductRepositoryException.java
@@ -2,7 +2,7 @@ package com.jpdr.apps.demo.webflux.stock.exception.product;
 
 public class ProductRepositoryException extends RuntimeException{
   
-  public ProductRepositoryException(Integer productId, Throwable ex){
+  public ProductRepositoryException(Long productId, Throwable ex){
     super("There was an error while retrieving the product " + productId, ex);
   }
