+++ src/main/java/com/jpdr/apps/demo/webflux/stock/exception/stock/InsufficientQuantityException.java
@@ -2,7 +2,7 @@ package com.jpdr.apps.demo.webflux.stock.exception.stock;
 
 public class InsufficientQuantityException extends RuntimeException{
   
-  public InsufficientQuantityException(Integer productId, Integer quantity){
+  public InsufficientQuantityException(Long productId, Integer quantity){
     super("The quantity of product "+ productId + " in stock is less than required. Current quantity: " + quantity);
   }
