+++ src/main/java/com/jpdr/apps/demo/webflux/stock/exception/stock/StockNotFoundException.java
@@ -2,7 +2,7 @@ package com.jpdr.apps.demo.webflux.stock.exception.stock;
 
 public class StockNotFoundException extends RuntimeException{
   
-  public StockNotFoundException(Integer productId){
+  public StockNotFoundException(Long productId){
     super("There is no stock for product " + productId);
   }
