+++ src/main/java/com/jpdr/apps/demo/webflux/stock/repository/stock/StockRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.reactive.ReactiveCrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface StockRepository extends ReactiveCrudRepository<Stock, Integer> {
+public interface StockRepository extends ReactiveCrudRepository<Stock, Long> {
 
 }
