+++ src/main/java/com/jpdr/apps/demo/webflux/stock/service/AppService.java
@@ -9,12 +9,12 @@ import java.util.List;
 public interface AppService {
   
   Mono<List<StockDto>> findAllStocks();
-  Mono<StockDto> findStockByProductId(Integer productId);
+  Mono<StockDto> findStockByProductId(Long productId);
   Mono<StockDto> createStock(StockDto stockDto);
   
   Mono<List<StockTransactionDto>> findAllTransactions();
-  Mono<List<StockTransactionDto>> findTransactions(Integer productId);
-  Mono<List<StockTransactionDto>> findTransactionsByProductId(Integer productId);
-  Mono<StockTransactionDto> createTransaction(Integer productId, StockTransactionDto transactionDto);
+  Mono<List<StockTransactionDto>> findTransactions(Long productId);
+  Mono<List<StockTransactionDto>> findTransactionsByProductId(Long productId);
+  Mono<StockTransactionDto> createTransaction(Long productId, StockTransactionDto transactionDto);
   
 }
